requirements.txt
Flask==2.0.2
Flask_Login==0.5.0
Flask_SQLAlchemy==2.5.1
Flask_WTF==1.0.0
mss==6.1.0
numpy==1.21.5
pandas==1.1.5
Pillow==9.0.0
pip==10.0.1
py2exe==0.11.0.1
requests==2.26.0
setuptools==39.0.1
SQLAlchemy==1.3.10
tkcalendar==1.6.1
ttkbootstrap==1.2.0
ttkwidgets==0.12.1
Werkzeug==2.0.2
WTForms==3.0.1
xlrd==2.0.1
